# backend/docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  handlers.CreateCustomerRequest:
    description: Datos para crear un nuevo cliente
    properties:
      documentNumber:
        example: "1234567890"
        type: string
      documentTypeId:
        example: 1
        type: integer
      email:
        example: maria.garcia@example.com
        type: string
      monthlyIncome:
        example: 5000000
        type: number
      name:
        example: María García
        type: string
      phoneNumber:
        example: +57 300 123 4567
        type: string
    type: object
...
  handlers.UpdateCustomerRequest:
    description: Datos para actualizar un cliente existente
    properties:
      documentNumber:
        example: "0987654321"
        type: string
      documentTypeId:
        example: 2
        type: integer
      email:
        example: maria.garcia.updated@example.com
        type: string
      monthlyIncome:
        example: 6000000
        type: number
      name:
        example: María García Actualizada
        type: string
      phoneNumber:
        example: +57 300 987 6543
        type: string
    type: object
...
  models.Customer:
    properties:
      createdAt:
        type: string
      createdById:
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      documentNumber:
        type: string
      documentTypeId:
        type: integer
      email:
        type: string
      id:
        type: integer
      monthlyIncome:
        type: number
      name:
        type: string
      phoneNumber:
        type: string
      status:
        type: boolean
      updatedAt:
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@example.com
    name: API Support
  description: API para gestión de solicitudes de crédito y evaluación de riesgo crediticio.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Credit Risk Assessment API
  version: "1.0"
paths:
...
  /customers:
    get:
      consumes:
      - application/json
      description: Retorna una lista de todos los clientes del sistema
      produces:
      - application/json
      responses:
        "200":
          description: Lista de clientes
          schema:
            items:
              $ref: '#/definitions/models.Customer'
            type: array
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Obtener todos los clientes
      tags:
      - Customers
    post:
      consumes:
      - application/json
      description: Crea un nuevo cliente en el sistema
      parameters:
      - description: Datos del cliente a crear
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Cliente creado exitosamente
          schema:
            $ref: '#/definitions/models.Customer'
        "400":
          description: Solicitud inválida
          schema:
            type: string
        "409":
          description: El cliente ya existe
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Crear un nuevo cliente
      tags:
      - Customers
  /customers/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un cliente del sistema
      parameters:
      - description: ID del cliente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Cliente eliminado exitosamente
        "400":
          description: ID inválido
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Eliminar un cliente
      tags:
      - Customers
    get:
      consumes:
      - application/json
      description: Retorna los detalles de un cliente específico
      parameters:
      - description: ID del cliente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cliente encontrado
          schema:
            $ref: '#/definitions/models.Customer'
        "400":
          description: ID inválido
          schema:
            type: string
        "404":
          description: Cliente no encontrado
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Obtener un cliente por ID
      tags:
      - Customers
    put:
      consumes:
      - application/json
      description: Actualiza los datos de un cliente existente
      parameters:
      - description: ID del cliente
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados del cliente
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cliente actualizado exitosamente
          schema:
            $ref: '#/definitions/models.Customer'
        "400":
          description: Solicitud inválida
          schema:
            type: string
        "404":
          description: Cliente no encontrado
          schema:
            type: string
        "409":
          description: El email ya existe
          schema:
            type: string
        "500":
          description: Error interno del servidor
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Actualizar un cliente
      tags:
      - Customers
...
securityDefinitions:
  BearerAuth:
    description: 'Ingresa el token JWT con el prefijo Bearer. Ejemplo: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"